Change adapter package name
