+++ app/src/main/java/com/samadihadis/androidappstore/peresentaion/adapters/boxStyle/AppBoxStyleViewHolder.kt
@@ -1,4 +1,4 @@
-package com.samadihadis.androidappstore.peresentaion.apps.boxStyle
+package com.samadihadis.androidappstore.peresentaion.adapters.boxStyle
 
 import android.view.View
 import androidx.appcompat.widget.AppCompatImageView
